Refactor MetadataValue::from_shared to use MetadataEncoding in a sane way
This is a step towards making this method do the right thing for
MetadataValue<Binary>.
