+++ src/metadata_encoding.rs
@@ -1,5 +1,15 @@
 use bytes::Bytes;
+use http::header::HeaderValue;
+use std::fmt;
+use std::error::Error;
 use std::hash::Hash;
+
+/// A possible error when converting a `MetadataValue` from a string or byte
+/// slice.
+#[derive(Debug, Default)]
+pub struct InvalidMetadataValue {
+    _priv: (),
+}
 
 // TODO(pgron): Make sealed
 pub trait ValueEncoding: Clone + Eq + PartialEq + Hash {
@@ -10,13 +20,15 @@ pub trait ValueEncoding: Clone + Eq + PartialEq + Hash {
     fn is_empty(value: &[u8]) -> bool;
 
     #[doc(hidden)]
-    fn encode(value: Bytes) -> Bytes;
+    fn from_shared(value: Bytes) -> Result<HeaderValue, InvalidMetadataValueBytes>;
 }
 
 #[derive(Clone, Debug, Eq, PartialEq, Hash)]
 pub struct Ascii {}
 #[derive(Clone, Debug, Eq, PartialEq, Hash)]
 pub struct Binary {}
+
+// ===== impl ValueEncoding =====
 
 impl ValueEncoding for Ascii {
     fn is_valid_key(key: &str) -> bool {
@@ -27,8 +39,11 @@ impl ValueEncoding for Ascii {
         value.is_empty()
     }
 
-    fn encode(value: Bytes) -> Bytes {
-       value
+    fn from_shared(value: Bytes) -> Result<HeaderValue, InvalidMetadataValueBytes> {
+       HeaderValue::from_shared(value)
+            .map_err(|_| {
+                InvalidMetadataValueBytes::new()
+            })
     }
 }
 
@@ -42,8 +57,56 @@ impl ValueEncoding for Binary {
         value.is_empty()
     }
 
-    fn encode(value: Bytes) -> Bytes {
+    fn from_shared(value: Bytes) -> Result<HeaderValue, InvalidMetadataValueBytes> {
         // TODO(pgron): Do this properly for base64
-        value
+       HeaderValue::from_shared(value)
+            .map_err(|_| {
+                InvalidMetadataValueBytes::new()
+            })
+    }
+}
+
+// ===== impl InvalidMetadataValue =====
+
+impl InvalidMetadataValue {
+    pub fn new() -> Self {
+        Default::default()
+    }
+}
+
+impl fmt::Display for InvalidMetadataValue {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        self.description().fmt(f)
+    }
+}
+
+impl Error for InvalidMetadataValue {
+    fn description(&self) -> &str {
+        "failed to parse metadata value"
+    }
+}
+
+/// A possible error when converting a `MetadataValue` from a string or byte
+/// slice.
+#[derive(Debug, Default)]
+pub struct InvalidMetadataValueBytes(InvalidMetadataValue);
+
+// ===== impl InvalidMetadataValueBytes =====
+
+impl InvalidMetadataValueBytes {
+    pub fn new() -> Self {
+        Default::default()
+    }
+}
+
+impl fmt::Display for InvalidMetadataValueBytes {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        self.0.fmt(f)
+    }
+}
+
+impl Error for InvalidMetadataValueBytes {
+    fn description(&self) -> &str {
+        self.0.description()
     }
 }
